requirements.txt
selenium
webdriver-manager
python-dotenv
pandas
deep_translator
vaderSentiment
langdetect
emoji